backend/requirements.txt
# Core FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Database and ORM
supabase==2.0.2
asyncpg==0.29.0
sqlalchemy==2.0.23
alembic==1.12.1

# AI and LangChain
langchain==0.0.350
langchain-openai==0.0.2
langgraph==0.0.20
openai==1.3.7
pydantic==2.5.0

# HTTP and async
httpx==0.25.2
aiohttp==3.9.1
aiofiles==23.2.1

# Data processing and analysis
pandas==2.1.4
numpy==1.25.2
scikit-learn==1.3.2
matplotlib==3.8.2
seaborn==0.13.0
python-dateutil==2.8.2
pytz==2023.3

# Image processing
pillow==10.1.0

# Utilities
redis==5.0.1
celery==5.3.4
structlog==23.2.0
loguru==0.7.2

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Security
cryptography==41.0.8
bcrypt==4.1.2

# Additional AI/ML libraries
transformers==4.36.0
torch==2.1.1
tensorflow==2.15.0
scipy==1.11.4

# Web scraping and parsing
beautifulsoup4==4.12.2
requests==2.31.0
lxml==4.9.3

# MQTT and messaging
asyncio-mqtt==0.16.1

# Email validation
email-validator==2.1.0
